Move query-parameter helper to separate file.
